% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/member_states_csts.R
\docType{data}
\name{member_states_csts}
\alias{member_states_csts}
\title{Cross-sectional time-series data on member states}
\format{
A data frame with 6 variables:
\describe{
\item{key_id}{\code{numeric}. An ID number that uniquely identifies each
observation.}
\item{year}{\code{numeric}. The year.}
\item{member_state_id}{\code{numeric}. An ID number that uniquely
identifies each member state.}
\item{member_state}{\code{string}. The name of the member sate.}
\item{member_state_code}{\code{string}. A two-letter code assigned to each
member state by the Commission.}
\item{years_as_member}{\code{numeric}. The number of years that the member
state has been a member, with the first year of membership coded \code{1}.}
}
}
\usage{
member_states_csts
}
\description{
This dataset is a template for cross-sectional time-series data on member
states. There is one observation per member state per year.
}
\keyword{datasets}
