% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/member_states_ddy.R
\docType{data}
\name{member_states_ddy}
\alias{member_states_ddy}
\title{Directed dyad-year template}
\format{
A data frame with 8 variables:
\describe{
\item{key_id}{Numeric. An ID number that uniquely identifies each
observation.}
\item{year}{Numeric. The year.}
\item{from_member_state_id}{Numeric. An ID number that uniquely identifies
each member state.}
\item{from_member_state}{String. The name of the first member state in the
directed dyad. }
\item{from_member_state_code}{String. A two-letter code assigned to each
member state by the Commission.}
\item{to_member_state_id}{Numeric. An ID number that uniquely identifies
each member state.}
\item{to_member_state}{String. The name of the second member state in the
directed dyad.}
\item{to_member_state_code}{String. A two-letter code assigned to each
member state by the Commission.}
}
}
\usage{
member_states_ddy
}
\description{
This dataset is a template for directed dyad-year data on member states.
There is one observation per directed dyad per year.
}
\keyword{datasets}
